fix: Fix KeyMod into u16

+++ src/event/keyboard/key_mod.rs
@@ -3,23 +3,23 @@ use bitflags::bitflags;
 use crate::bind;
 
 bitflags! {
-  pub struct KeyMod : u32 {
-    const NONE = bind::SDL_Keymod_KMOD_NONE;
-    const LSHIFT = bind::SDL_Keymod_KMOD_LSHIFT;
-    const RSHIFT = bind::SDL_Keymod_KMOD_RSHIFT;
-    const LCTRL = bind::SDL_Keymod_KMOD_LCTRL;
-    const RCTRL = bind::SDL_Keymod_KMOD_RCTRL;
-    const LALT = bind::SDL_Keymod_KMOD_LALT;
-    const RALT = bind::SDL_Keymod_KMOD_RALT;
-    const LGUI = bind::SDL_Keymod_KMOD_LGUI;
-    const RGUI = bind::SDL_Keymod_KMOD_RGUI;
-    const NUM = bind::SDL_Keymod_KMOD_NUM;
-    const CAPS = bind::SDL_Keymod_KMOD_CAPS;
-    const MODE = bind::SDL_Keymod_KMOD_MODE;
-    const RESERVED = bind::SDL_Keymod_KMOD_RESERVED;
-    const CTRL = bind::SDL_Keymod_KMOD_CTRL;
-    const SHIFT = bind::SDL_Keymod_KMOD_SHIFT;
-    const ALT = bind::SDL_Keymod_KMOD_ALT;
-    const GUI = bind::SDL_Keymod_KMOD_GUI;
+  pub struct KeyMod : u16 {
+    const NONE = bind::SDL_Keymod_KMOD_NONE as u16;
+    const LSHIFT = bind::SDL_Keymod_KMOD_LSHIFT as u16;
+    const RSHIFT = bind::SDL_Keymod_KMOD_RSHIFT as u16;
+    const LCTRL = bind::SDL_Keymod_KMOD_LCTRL as u16;
+    const RCTRL = bind::SDL_Keymod_KMOD_RCTRL as u16;
+    const LALT = bind::SDL_Keymod_KMOD_LALT as u16;
+    const RALT = bind::SDL_Keymod_KMOD_RALT as u16;
+    const LGUI = bind::SDL_Keymod_KMOD_LGUI as u16;
+    const RGUI = bind::SDL_Keymod_KMOD_RGUI as u16;
+    const NUM = bind::SDL_Keymod_KMOD_NUM as u16;
+    const CAPS = bind::SDL_Keymod_KMOD_CAPS as u16;
+    const MODE = bind::SDL_Keymod_KMOD_MODE as u16;
+    const RESERVED = bind::SDL_Keymod_KMOD_RESERVED as u16;
+    const CTRL = bind::SDL_Keymod_KMOD_CTRL as u16;
+    const SHIFT = bind::SDL_Keymod_KMOD_SHIFT as u16;
+    const ALT = bind::SDL_Keymod_KMOD_ALT as u16;
+    const GUI = bind::SDL_Keymod_KMOD_GUI as u16;
   }
 }
